requirements.txt
# requirements.txt
httpx==0.27.0
python-dotenv==1.0.1
requests==2.31.0
opencv-python
# Only install on Linux (e.g., Raspberry Pi)
picamera2; sys_platform == "linux"

# 